Changed the 3 line X, Y, Z fields to two seperate text fields with new layouts for formatting purposes.
The change was necessary because that the X Y Z 'header' which identified each value was being shifted all over the view as the length of the value's numeral changed. It made the data difficult to read. Accelerometer and gyroscope acceleration did not have this issue as the values don't go above 1 numeral place, but I changed them as well for consistency through the app.